requirements.txt
transformers>=4.42.0
sentencepiece>=0.1.99
huggingface-hub>=0.23.0
accelerate>=0.33.0
peft>=0.12.0
datasets>=2.20.0
rouge-score>=0.1.2
scikit-learn>=1.5.1
pywin32>=306

# Install PyTorch CPU wheels on Windows:
# pip install torch --index-url https://download.pytorch.org/whl/cpu

pypdf>=4.2.0
pdfminer.six>=20231228
python-docx>=1.1.2

numpy>=1.26.4
regex>=2024.5.15
tqdm>=4.66.4

# Gemini client (kept optional for prior flow)
google-generativeai>=0.7.2
requests>=2.32.3
python-dotenv>=1.0.1
dotenv